+++ src/level.cpp
@@ -1,192 +1,12 @@
 #include "level.h"
 
-// LEVELS
-
-// PRIVATE
-void Level::Levels::getMenuInfos(LevelData* data, json* level) {
-    json menuInfos = (*level)["menuInfos"];
-    json meColor = menuInfos["backgroundColor"];
-
-    data->name = menuInfos["name"];
-    data->difficulty = menuInfos["difficulty"];
-    data->menuColor = { static_cast<Uint8>(meColor["r"]), static_cast<Uint8>(meColor["g"]), static_cast<Uint8>(meColor["b"]), 255 };
-}
-
-void Level::Levels::getConfigInfos(LevelData* data, json* level)  {
-    json config = (*level)["config"];
-
-    json bgCol = config["background"]["color"];
-    data->backgroundColor = { static_cast<Uint8>(bgCol["r"]), static_cast<Uint8>(bgCol["g"]), static_cast<Uint8>(bgCol["b"]), 255 };
-    json grCol = config["ground"]["color"];
-    data->groundColor = { static_cast<Uint8>(grCol["r"]), static_cast<Uint8>(grCol["g"]), static_cast<Uint8>(grCol["b"]), 255 };
-    json liCol = config["line"]["color"];
-    data->lineColor = { static_cast<Uint8>(liCol["r"]), static_cast<Uint8>(liCol["g"]), static_cast<Uint8>(liCol["b"]), 255 };
-
-    data->backgroundTexture = config["background"]["texture"];
-    data->groundTexture = config["ground"]["texture"];
-    data->lineTexture = config["line"]["texture"];
-
-    
-    for (int i = 0; i < config["colors"].size(); i++) {
-        json col = config["colors"][i];
-        data->colors.push_back({ static_cast<Uint8>(col["r"]), static_cast<Uint8>(col["g"]), static_cast<Uint8>(col["b"]), 255 });
-    }
-
-    data->song = config["song"];
-}
-
-void Level::Levels::getBlocksInfos(LevelData* data, json* level, json* itemsData)  {
-    for (json block : (*level)["blocks"]) {
-        json itemData = (*itemsData)[block["i"]];
-        BufferedBlock* bBlock = new BufferedBlock();
-        bBlock->type = itemData["type"];
-
-        bBlock->texture = std::string(block["i"]) + ".png";
-
-        json pos = block["p"];
-        bBlock->pos = { pos["x"], pos["y"] };
-        json texSize = itemData["size"]["texture"];
-        bBlock->textureSize = { texSize["w"], texSize["h"] };
-        json texOffset = itemData["offset"]["texture"];
-        bBlock->textureOffset = { texOffset["x"], texOffset["y"] };
-        
-        if (block.contains("r")) bBlock->rotation = block["r"];
-        json hitSize = itemData["size"]["hitbox"];
-        json hitOffset = itemData["offset"]["hitbox"];
-        switch (static_cast<int>(bBlock->rotation)) {
-            case 90:
-                bBlock->hitboxSize = { hitSize["h"], hitSize["w"] };
-                bBlock->hitboxOffset = { hitOffset["y"], hitOffset["x"] };
-                break;
-            case 180:
-                bBlock->hitboxSize = { hitSize["w"], hitSize["h"] };
-                bBlock->hitboxOffset = { 1.0f - static_cast<float>(hitSize["w"]) - static_cast<float>(hitOffset["x"]), 1.0f - static_cast<float>(hitSize["h"]) - static_cast<float>(hitOffset["y"]) };
-                break;
-            case 270:
-                bBlock->hitboxSize = { hitSize["h"], hitSize["w"] };
-                bBlock->hitboxOffset = { 1.0f - static_cast<float>(hitSize["h"]) - static_cast<float>(hitOffset["y"]), 1.0f - static_cast<float>(hitSize["w"]) - static_cast<float>(hitOffset["x"]) };
-                break;
-            default:
-                bBlock->hitboxSize = { hitSize["w"], hitSize["h"] };
-                bBlock->hitboxOffset = { hitOffset["x"], hitOffset["y"] };
-                break;
-        }
-
-        if (block.contains("c")) bBlock->colorID = block["c"];
-
-        if (block.contains("f")) {
-            bool fx = (std::string(block["f"]).find("x") != std::string::npos);
-            bool fy = (std::string(block["f"]).find("y") != std::string::npos);
-
-            bBlock->flip = (fx && fy) ? SDL_FLIP_NONE : (fx) ? SDL_FLIP_HORIZONTAL : (fy) ? SDL_FLIP_VERTICAL : SDL_FLIP_NONE;
-        } else bBlock->flip = SDL_FLIP_NONE;
-
-        bBlock->zIndex = new Zindex(Zindex::getLayer(itemData["zIndex"]["layer"]), itemData["zIndex"]["order"]);
-        if (itemData.contains("special")) {
-            bBlock->specialData = ItemSpecialData();
-            bBlock->specialData.type = itemData["special"]["type"];
-            bBlock->specialData.desc = itemData["special"]["desc"];
-        }
-
-        data->blocks.push_back(bBlock);
-    }
-}
-
-void Level::Levels::getTriggersInfos(LevelData* data, json* level, json* itemsData) {
-    for (json trigger : (*level)["triggers"]) {
-        BufferedTrigger* bTrigger = new BufferedTrigger();
-        json itemData = (*itemsData)[trigger["i"]];
-
-        if (itemData["type"] != STARTPOS) {
-            bTrigger->type = itemData["type"];
-            bTrigger->pos = { trigger["p"]["x"], trigger["p"]["y"] };
-            bTrigger->color = { static_cast<Uint8>(trigger["c"]["r"]), static_cast<Uint8>(trigger["c"]["g"]), static_cast<Uint8>(trigger["c"]["b"]) };
-            bTrigger->ms = trigger["d"];
-            if (trigger.contains("tt")) bTrigger->touchTrigger = trigger["tt"];
-
-            data->triggers.push_back(bTrigger);
-        }
-    }
-}
-
-void Level::Levels::getStartposInfo(LevelData* data, json* level, json* itemsData) {
-    static GameData* gameData = new GameData();
-    json config = (*level)["config"];
-    Checkpoint* startpos = nullptr;
-
-    for (json trigger : (*level)["triggers"]) if ((*itemsData)[trigger["i"]]["type"] == STARTPOS && trigger["p"]["x"] >= 0) {
-        if (!startpos) startpos = new Checkpoint();
-        
-        if (startpos->playerPos.x <= trigger["p"]["x"]) {
-            startpos->camPos = { (float)(trigger["p"]["x"]) + gameData->positions->camera.x, (float)(trigger["p"]["y"]) - 10.0f + gameData->sizes->cameraPaddingTop};
-            startpos->playerPos = { trigger["p"]["x"], trigger["p"]["y"] };
-            startpos->velocity = { gameData->physics->speeds[trigger["sp"]], 0 };
-            startpos->speed = trigger["sp"];
-            startpos->gravity = trigger["gr"];
-            startpos->size = trigger["si"];
-            startpos->gamemode = trigger["ga"];
-        }
-    }
-
-    if (!startpos) {
-        startpos = new Checkpoint();
-        startpos->camPos = gameData->positions->camera;
-        startpos->velocity = { gameData->physics->speeds[config["speed"]], 0 };
-        startpos->speed = config["speed"];
-        startpos->gravity = config["gravity"];
-        startpos->size = config["size"];
-        startpos->gamemode = config["gamemode"];
-    }
-
-    data->startpos = *startpos;
-}
-
-// PUBLIC
-LevelData* Level::Levels::getLevelData(int id) {
-    Uint32 start = SDL_GetTicks();
-    std::cout << "Loading level data..." << std::endl;
-
-    LevelData* data = new LevelData();
-    json* levels = H2DE_Json::read("data/levels.json");
-
-    std::string stingifiedID = std::to_string(id);
-    if ((*levels).contains(stingifiedID)) {
-
-        json* level = &(*levels)[std::to_string(id)];
-        try {
-            data->id = id;
-            getMenuInfos(data, level);
-            getConfigInfos(data, level);
-
-            json* itemsData = H2DE_Json::read("data/items.json");
-            getBlocksInfos(data, level, itemsData);
-            getTriggersInfos(data, level, itemsData);
-            getStartposInfo(data, level, itemsData);
-
-        } catch (const std::exception& e) {
-            throw std::runtime_error("HGD-1004: Error loading level => " + std::string(e.what()));
-        }
-    } else throw std::runtime_error("HGD-1005: Error loading level => Level ID not found");
-
-    std::cout << "Level data loaded in " << SDL_GetTicks() - start << "ms" << std::endl;
-    return data;
-}
-
-void Level::Levels::loadLevelData(int id, const fs::path& gmdFile) {
-
-}
-
-
-
-// LEVEL
-
 // INIT
 Level::Level(Game* g, int i) : game(g), id(i) {
     static H2DE_Engine* engine = game->getEngine();    
 
     std::cout << "Initializing level..." << std::endl;
     Uint32 start = SDL_GetTicks();
-    data = Levels::getLevelData(id);
+    data = LevelLoader::getLevelData(id);
 
     initAttempts();
     initCamera();
@@ -292,7 +112,7 @@ void Level::update() {
 
     switch (state.main) {
         case LEVEL_PLAYING:
-            backgroundPos.x += (gameData->physics->speeds[speed] * gameData->physics->backgroundRatio);
+            updateBackground();
             H2DE_TickTimelineManager(topGroundTM);
             H2DE_TickTimelineManager(botGroundTM);
             for (Item* item : items) item->update();
@@ -300,6 +120,21 @@ void Level::update() {
             camera->update();
             break;
     }
+}
+
+void Level::updateBackground() {
+    static GameData* gameData = game->getData();
+    static Camera* camera = game->getCamera();
+
+    static float camMinY = gameData->positions->cameraMinY;
+    static float camMaxY = gameData->positions->cameraMaxY;
+    static float bgMinY = gameData->positions->backgroundMinY;
+    static float bgMaxY = gameData->positions->backgroundMaxY;
+
+    float camPosY = camera->getPos().y;
+
+    backgroundPos.x += (gameData->physics->speeds[speed] * gameData->physics->backgroundRatio);
+    backgroundPos.y = ((camPosY - camMinY) / (camMaxY - camMinY)) * (bgMaxY - bgMinY) + bgMinY;
 }
 
 // RENDER
@@ -386,35 +221,31 @@ bool Level::sortItems(Item* item1, Item* item2) {
 }
 
 // GETTER
-LevelData* Level::getData() {
+LevelData* Level::getData() const {
     return data;
 }
 
-int Level::getCurrentSpeed() {
-    return speed;
-}
-
-LevelPos Level::getBotGroundPos() {
+LevelPos Level::getBotGroundPos() const {
     return botGroundPos;
 }
 
-LevelPos Level::getTopGroundPos() {
+LevelPos Level::getTopGroundPos() const {
     return topGroundPos;
 }
 
-Color Level::getBackgroundColor() {
+Color Level::getBackgroundColor() const {
     return backgroundColor;
 }
 
-Color Level::getGroundColor() {
+Color Level::getGroundColor() const {
     return groundColor;
 }
 
-Color Level::getLineColor() {
+Color Level::getLineColor() const {
     return lineColor;
 }
 
-Player* Level::getPlayer() {
+Player* Level::getPlayer() const {
     return player;
 }
 
@@ -422,7 +253,11 @@ std::vector<Item*>* Level::getItems() {
     return &items;
 }
 
-int Level::getLevelLength() {
+int Level::getCurrentSpeed() const {
+    return speed;
+}
+
+int Level::getLevelLength() const {
     return levelLength;
 }